refactor(agents): implement dedicated langgraph error handling module
