+++ src/agents/builtin/lib.rs
@@ -84,6 +84,7 @@ pub mod agentic_seek;
 pub mod pi;
 pub mod ruflo_plugins;
 pub mod tool_executor_engine;
+pub mod error_handling;
 
 fn get_env(key: &str, default: &str) -> String {
     std::env::var(key).unwrap_or_else(|_| default.to_string())
